news-rss-feed-sitrep/requirements.txt
# OpenCTI RSS SITREP Connector v2 — Python dependencies
#
# All ML capabilities run through the PyTorch + Hugging Face Transformers stack.
# No spaCy, no NLTK, no sumy — see DESIGN.md for rationale.

# ── OpenCTI SDK ────────────────────────────────────────────────────────────────
pycti==7.260309.0.3
stix2>=3.0.1

# ── ML backbone ───────────────────────────────────────────────────────────────
torch                          # required by all Transformers-based models
transformers>=4.40.0           # zero-shot classifier (DeBERTa) + optional DistilBART
sentence-transformers>=2.7.0   # MiniLM embeddings (thin wrapper over Transformers)
gliner>=0.2.0                  # zero-shot NER; entity types defined at inference time

# ── RSS and web scraping ──────────────────────────────────────────────────────
feedparser>=6.0.10             # RSS/Atom feed parsing
trafilatura>=1.8.0             # inline full-text article extraction
curl_cffi>=0.7.0               # Chrome TLS impersonation for 403-resistant scraping

# ── Utilities ─────────────────────────────────────────────────────────────────
requests>=2.31.0               # fallback HTTP (curl_cffi preferred)
beautifulsoup4>=4.12.0         # HTML cleaning fallback
pyyaml>=6.0
numpy                          # comes with torch; listed explicitly for clarity
